[SWE API] Initial support for observation statistics

+++ sensorhub-core/src/main/java/org/sensorhub/api/datastore/obs/ObsStatsQuery.java
@@ -144,7 +144,7 @@ public class ObsStatsQuery implements IQueryFilter
         }      
 
 
-        public B withLimit(int limit)
+        public B withLimit(long limit)
         {
             instance.limit = limit;
             return (B)this;
